Update CDVInAppBrowser.m
Alteration to the logic of handling a url that has a target of _system. Is now as it was at 1.4.0 release where it will ask the system to open the URL and if the system fails will fall back to calling a plugin. As iOS 10 has depreciated openURL function in favour of openURL with a completion handler, the code detects the support of the new openURL function and uses it is available. If it isn't it falls back to the old openURL function without the completion handler.